+++ experimental/packages/opentelemetry-exporter-metrics-otlp-grpc/test/OTLPMetricExporter.test.ts
@@ -174,13 +174,16 @@ const testOTLPMetricExporter = (params: TestParams) => {
         const certPaths = [
           './test/certs/ca.crt',
           './test/certs/client.crt',
-          './test/certs/server.crt'
+          './test/certs/server.crt',
         ];
         certPaths.forEach(certPath => {
           const cert = new X509Certificate(fs.readFileSync(certPath));
           const now = new Date();
-          assert.ok(new Date(cert.validTo) > now, `TLS cert "${certPath}" is still valid: cert.validTo="${cert.validTo}" (if this fails use 'npm run maint:regenerate-test-certs')`);
-        })
+          assert.ok(
+            new Date(cert.validTo) > now,
+            `TLS cert "${certPath}" is still valid: cert.validTo="${cert.validTo}" (if this fails use 'npm run maint:regenerate-test-certs')`
+          );
+        });
       });
     }